Fixed warning on DEC

+++ mmm2d.c
@@ -910,7 +910,7 @@ static char *MMM2D_tune_near(double error)
 
   realloc_intlist(&besselCutoff, besselCutoff.n = P);
   for (p = 1; p < P; p++)
-    besselCutoff.e[p-1] = (int)floor(P/(2*p)) + 1;
+    besselCutoff.e[p-1] = (int)floor(((double)P)/(2*p)) + 1;
 
   /* complex sum, determine cutoffs (dist dependent) */
   T = log(part_error/(16*M_SQRT2)*box_l[1]*box_l[2]);
